Fixes and ID as Long instead of Integer

+++ src/main/java/com/sportmonks/data/entity/FixtureEvent.java
@@ -16,19 +16,19 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 public class FixtureEvent {
 
 	@JsonProperty("id")
-	private Integer id;
+	private Long id;
 	@JsonProperty("team_id")
 	private String teamId;
 	@JsonProperty("type")
 	private String type;
 	@JsonProperty("fixture_id")
-	private Integer fixtureId;
+	private Long fixtureId;
 	@JsonProperty("player_id")
-	private Integer playerId;
+	private Long playerId;
 	@JsonProperty("player_name")
 	private String playerName;
 	@JsonProperty("related_player_id")
-	private Integer relatedPlayerId;
+	private Long relatedPlayerId;
 	@JsonProperty("related_player_name")
 	private String relatedPlayerName;
 	@JsonProperty("minute")
@@ -43,12 +43,12 @@ public class FixtureEvent {
 	private Map<String, Object> additionalProperties = new HashMap<String, Object>();
 
 	@JsonProperty("id")
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
 	@JsonProperty("id")
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 
@@ -73,22 +73,22 @@ public class FixtureEvent {
 	}
 
 	@JsonProperty("fixture_id")
-	public Integer getFixtureId() {
+	public Long getFixtureId() {
 		return fixtureId;
 	}
 
 	@JsonProperty("fixture_id")
-	public void setFixtureId(Integer fixtureId) {
+	public void setFixtureId(Long fixtureId) {
 		this.fixtureId = fixtureId;
 	}
 
 	@JsonProperty("player_id")
-	public Integer getPlayerId() {
+	public Long getPlayerId() {
 		return playerId;
 	}
 
 	@JsonProperty("player_id")
-	public void setPlayerId(Integer playerId) {
+	public void setPlayerId(Long playerId) {
 		this.playerId = playerId;
 	}
 
@@ -103,12 +103,12 @@ public class FixtureEvent {
 	}
 
 	@JsonProperty("related_player_id")
-	public Integer getRelatedPlayerId() {
+	public Long getRelatedPlayerId() {
 		return relatedPlayerId;
 	}
 
 	@JsonProperty("related_player_id")
-	public void setRelatedPlayerId(Integer relatedPlayerId) {
+	public void setRelatedPlayerId(Long relatedPlayerId) {
 		this.relatedPlayerId = relatedPlayerId;
 	}
